requirements.txt
# Co-DETR PyTorch Implementation - Minimal Dependencies
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Numerical Operations
numpy>=1.24.0

# Image I/O and Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Configuration Management
PyYAML>=6.0

# Progress Bars and Utilities
tqdm>=4.65.0

# Training Visualization (Optional)
tensorboard>=2.13.0

# Development Tools (Optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
